separate the seeders with version
