+++ database/seeders/v2_0_0/BussinessSeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Database\Seeders;
+namespace Database\Seeders\v2_0_0;
 
 use App\Model\Common\Bussiness;
 use Illuminate\Database\Seeder;
