+++ database/seeders/v2_0_0/RazorpaySupportedCurrencySeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace database\seeds;
+namespace database\seeds\v2_0_0;
 
 use App\Plugins\Razorpay\Model\RazorpayPayment;
 use Database\Seeders\RazorpaySupportedCurrencySeeder;
